Create Kate and her 7 old sister
Kate and her 7-year-old sister Tina were playing a spelling game. Kate will tell a number and
Tina has to find the corresponding month and should spell the month. Can you write a C
program to automate it? Accept a month in digit from the user. Display the corresponding
month in words. If the number is not between 1 and 12 display the message “Invalid Month”.